Restruct directories (#3)
