+++ .gitignore
@@ -7,7 +7,7 @@
 /coverage
 
 # production
-/build
+/lib
 
 # misc
 .DS_Store
