+++ config-overrides/index.build.js
@@ -1,1 +1,37 @@
-module.exports = require('../lib/config-overrides');
+const { compose } = require('react-app-rewired');
+const rewireEslint = require('react-app-rewire-eslint');
+const rewireReactLibrary = require('react-app-rewire-react-library');
+const path = require('path');
+const { initPaths, replacePaths } = require('./utils');
+const pkg = require('../package.json');
+
+const appDirectory = path.resolve(__dirname, '..');
+const resolveOwn = relativePath => path.resolve(appDirectory, relativePath);
+
+initPaths(appDirectory);
+
+replacePaths({
+  appBuild: resolveOwn('lib'),
+  appPublic: resolveOwn('lib'),
+  appHtml: resolveOwn(''),
+});
+
+module.exports = {
+  webpack: (config, env) => {
+    const rewires = compose(
+      rewireEslint,
+      (config, env) => {
+        const options = {
+          ...pkg,
+          module: resolveOwn(pkg.module),
+          main: resolveOwn(pkg.main),
+        };
+        config = rewireReactLibrary(config, env, options, true);
+        config.entry = options.module;
+        config.output.library = undefined;
+        return config;
+      },
+    );
+    return rewires(config, env);
+  },
+};
