+++ example/src/index.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import { Menu } from '@ridi/cms-ui';
+import { Menu } from '../../src';
 
 ReactDOM.render(<Menu />, document.getElementById('root'));
