--- lib/package.json
@@ -1,20 +1,0 @@
-{
-  "name": "@ridi/cms-ui",
-  "version": "0.1.0",
-  "module": "./src/index.js",
-  "main": "./lib/cms-ui.js",
-  "scripts": {
-    "build": "react-app-rewired build"
-  },
-  "dependencies": {
-    "react": "^16.2.0",
-    "react-dom": "^16.2.0"
-  },
-  "devDependencies": {
-    "@ridi/eslint-config": "^4.0.0",
-    "react-app-rewire-eslint": "^0.2.3",
-    "react-app-rewire-react-library": "^0.1.1",
-    "react-app-rewired": "^1.5.0",
-    "react-scripts": "1.1.1"
-  }
-}
